finish renaming information component, switch unavailable to available in selectors, clarify function of balance chart selector
